make_get_request вынес в фикструты приложения fastapi.
